# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PROBOX Documentación",
        "description": "Adminstración de Proyectos APP",
        "contact": {
            "email": "daruiza@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8090/api",
            "description": "PROBOX API Server"
        }
    ],
    "paths": {
...
        "/task/index": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Get All Task",
                "description": "Return Task",
                "operationId": "getTask",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/task/store": {
            "post": {
                "tags": [
                    "Task"
                ],
                "summary": "Store Task",
                "description": "Store Task",
                "operationId": "storeTask",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Task"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/task/showbyid/{id}": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Get One Task By one Id",
                "description": "Return One Task",
                "operationId": "getTaskById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/task/update/{id}": {
            "put": {
                "tags": [
                    "Task"
                ],
                "summary": "Update One Task By one Id",
                "description": "Update One Task",
                "operationId": "getUpdateTaskById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Task"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/task/destroy/{id}": {
            "delete": {
                "tags": [
                    "Task"
                ],
                "summary": "Delete One Task By one Id",
                "description": "Delete One Task",
                "operationId": "getDestroyTaskById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Task Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/task/showevidencebyid/{id}": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Get One Evidence By one Id",
                "description": "Return One Evidence by Id",
                "operationId": "getShowEvidenceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Show Evidence by Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/task/showprojectbyid/{id}": {
            "get": {
                "tags": [
                    "Task"
                ],
                "summary": "Get One Project By one Id",
                "description": "Return One Project by Id",
                "operationId": "getShowProjectById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Show Project by Id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Task": {
                "title": "Task",
                "description": "Task body data",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the Task",
                        "type": "string",
                        "example": "Name Task"
                    },
                    "description": {
                        "title": "description",
                        "description": "Description of the Task",
                        "type": "string",
                        "example": "Description Task"
                    },
                    "date_init": {
                        "title": "date_init",
                        "description": "Date_init of the Task",
                        "type": "string",
                        "format": "date",
                        "example": "Date_init Task"
                    },
                    "date_closed": {
                        "title": "date_closed",
                        "description": "Date_closed of the Task",
                        "type": "string",
                        "format": "date",
                        "example": "Date_closed Task"
                    },
                    "focus": {
                        "title": "focus",
                        "description": "Focus of the Task",
                        "type": "boolean",
                        "example": "Focus Task"
                    },
                    "id_task": {
                        "title": "id_task",
                        "description": "Id_task of the Task",
                        "example": "Id_task Task"
                    },
                    "project_id": {
                        "title": "project_id",
                        "description": "Project_id of the Task",
                        "example": "Project_id Task"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}